s2: Allow CustomEncoder fallback
If a custom encoder returns a negative value to the compression with regular methods.

+ Split stateful reader+writer code into separate files.
